Refactor calibration: extension, Monad stats, docs
- Move pyabc-dependent code to PCMMCalibrationExt (PyCall weak dep)
- Add finalPopulationCount(Monad) averaging replicates; error on empty
- Add meanPopulationTimeSeries built-in summary statistic
- Export posterior; fix loading-order docs (order doesn't matter)
- CalibrationTests: use Base.get_extension for extension internals
