+++ src/calibration/calibration.jl
@@ -1,4 +1,4 @@
-using PyCall, Dates, CSV, Statistics
+using Dates, CSV, Statistics
 
 include("problem.jl")
 include("distance.jl")
@@ -73,50 +73,3 @@ function calibrationMonadIDs(calibration::Calibration)
     return constituentIDs(path_to_csv)
 end
 
-################## Python Environment ##################
-
-"""
-    _importPyABC()
-
-Import the `pyabc` Python module, using the UQ-specific Python environment
-(`PCMM_UQ_PYTHON_PATH` / `pcmm_globals.path_to_uq_python`) when set.
-
-Raises a helpful error if `pyabc` is not importable.
-"""
-function _importPyABC()
-    uq_python = pcmm_globals.path_to_uq_python
-    if !ismissing(uq_python)
-        # Ensure PyCall uses the configured Python; no-op if already configured
-        if PyCall.python != uq_python
-            @warn """
-            PyCall is currently using $(PyCall.python) but PCMM_UQ_PYTHON_PATH is set to
-            $(uq_python). PyCall cannot switch Python interpreters at runtime.
-            To use the UQ Python environment, rebuild PyCall before starting Julia:
-
-                ENV["PYTHON"] = "$(uq_python)"
-                import Pkg; Pkg.build("PyCall")
-            """
-        end
-    end
-
-    try
-        return pyimport("pyabc")
-    catch
-        throw(ErrorException("""
-        Could not import `pyabc`. Make sure it is installed in the Python environment
-        that PyCall is using (currently: $(PyCall.python)).
-
-        Setup instructions:
-            conda create -n pcmm-uq python=3.10
-            conda activate pcmm-uq
-            pip install pyabc
-
-        Then point PCMM to this environment (add to your shell profile or set per-session):
-            export PCMM_UQ_PYTHON_PATH="/path/to/conda/envs/pcmm-uq/bin/python"
-
-        Then rebuild PyCall in Julia (once per machine):
-            ENV["PYTHON"] = ENV["PCMM_UQ_PYTHON_PATH"]
-            import Pkg; Pkg.build("PyCall")
-        """))
-    end
-end
